# api-gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Second Opinion Platform API Gateway
  description: |
    API Gateway for the Second Opinion Platform microservices architecture.
    
    This gateway provides unified access to all platform services including:
    - Patient Identity Service (Authentication & Authorization)
    - Case Management Service (Medical cases & documents)
    - AI Analysis Service (Medical AI insights)
    - Professional Service (Medical professionals)
    - Notification Service (Multi-channel communications)
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication obtained from the identity service.
    
    ## Rate Limiting
    API endpoints are rate limited based on user type and endpoint category.
    
    ## Error Handling
    All responses follow a consistent error format with appropriate HTTP status codes.
    
  version: 1.0.0
  contact:
    name: Second Opinion Platform Team
    email: support@secondopinion.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.secondopinion.com
    description: Production server
paths:
...
  /api/v1/cases:
    get:
      summary: List cases
      description: Retrieve list of medical cases
      tags:
        - Cases
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SUBMITTED, PROFESSIONAL_REVIEWING, COMPLETED]
      responses:
        '200':
          description: Cases retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseList'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    post:
      summary: Create case
      description: Create a new medical case
      tags:
        - Cases
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCase'
      responses:
        '201':
          description: Case created successfully
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateCase:
      type: object
      required:
        - title
        - description
        - category
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [GENERAL, CARDIOLOGY, NEUROLOGY, ONCOLOGY, ORTHOPEDIC]
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
          default: NORMAL
        chiefComplaint:
          type: string
    CaseList:
      type: object
      properties:
        success:
          type: boolean
        cases:
          type: array
          items:
            $ref: '#/components/schemas/Case'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            pages:
              type: integer
    Case:
      type: object
      properties:
        id:
          type: string
        caseNumber:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...